+++ test/equal_test.dart
@@ -4,7 +4,7 @@ import 'package:callor/callor.dart';
 void main() {
   const error = 'different value';
   const value = 'an input value';
-  const equal = Equal(value, diff: error);
+  final equal = Equal(() => value, diff: error);
   group('Equal validator', () {
     test('equivalent values', () {
       expect(equal(value), null);
@@ -13,6 +13,11 @@ void main() {
     test('different values', () {
       expect(equal(error), error);
       expect(equal.to(error), error);
+    });
+    test('null', () {
+      final nullEqual = Equal(() => null, diff: error);
+      expect(nullEqual(null), null);
+      expect(nullEqual.to(null), null);
     });
   });
 }